Refactor to response class to be in their own files.
