+++ lib/convertkit/resources/tags.rb
@@ -40,7 +40,7 @@ module ConvertKit
 
         response = @client.post("#{PATH}/#{tag_id}/subscribe", request)
 
-        SubscriptionResponse.new(response)
+        TagSubscriptionResponse.new(response)
       end
 
       def remove_tag_from_subscriber(tag_id, subscriber_id)
@@ -53,42 +53,6 @@ module ConvertKit
         response = @client.post("#{PATH}/#{tag_id}/unsubscribe", { email: subscriber_email})
 
         TagResponse.new(response)
-      end
-    end
-
-    class TagResponse
-      attr_accessor :id, :name, :account_id, :state, :created_at, :updated_at, :deleted_at
-
-      def initialize(response)
-        @id = response['id']
-        @name = response['name']
-        @account_id = response['account_id']
-        @state = response['state']
-        @created_at = DateTime.parse(response['created_at']) unless response.fetch('created_at', '').strip.empty?
-        @updated_at = DateTime.parse(response['updated_at']) unless response.fetch('updated_at', '').strip.empty?
-        @deleted_at = DateTime.parse(response['deleted_at']) unless response.fetch('deleted_at', '').strip.empty?
-      end
-    end
-
-    class SubscriptionResponse
-      attr_accessor :id,
-                    :state,
-                    :source,
-                    :referrer,
-                    :subscribable_id,
-                    :subscribable_type,
-                    :subscriber_id,
-                    :created_at
-
-      def initialize(response)
-        @id = response['id']
-        @state = response['state']
-        @source = response['source']
-        @referrer = response['referrer']
-        @subscribable_id = response['subscribable_id']
-        @subscribable_type = response['subscribable_type']
-        @subscriber_id = response.dig('subscriber', 'id')
-        @created_at = DateTime.parse(response['created_at']) unless response.fetch('created_at', '').strip.empty?
       end
     end
   end
